requirements.txt
# AI 客服語音克隆系統 - 完整依賴套件

# Flask 核心
Flask==3.1.0
Flask-Cors==5.0.0
Werkzeug==3.1.3
Jinja2==3.1.4
MarkupSafe==3.0.2
click==8.1.7
itsdangerous==2.2.0
blinker==1.9.0

# JWT 認證
Flask-JWT-Extended==4.6.0
PyJWT==2.9.0

# 即時通訊（WebSocket）- 新增
Flask-SocketIO==5.5.1
python-socketio==5.14.3
python-engineio==4.12.3
simple-websocket==1.1.0
wsproto==1.3.1
h11==0.16.0
bidict==0.23.1

# Qwen API 客戶端 - 新增
dashscope==1.25.1

# WebSocket 客戶端 - 新增
websocket-client==1.9.0

# Excel 匯出功能
openpyxl==3.1.5

# HTTP 請求
requests==2.32.3
certifi==2024.8.30
charset-normalizer==3.4.0
urllib3==2.2.3
idna==3.10

# AI/LLM 相關
openai==1.54.3

# 語音處理
SpeechRecognition==3.10.4

# 音頻處理
pydub==0.25.1

# 環境變數管理
python-dotenv==1.0.1

# 基礎工具
colorama==0.4.6

# EPUB 解析（AI 廣播劇）
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0

# 數據處理
numpy==1.26.4

# 簡繁轉換
opencc-python-reimplemented==0.1.7
